+++ src/JsonApi/Request/Request.php
@@ -105,7 +105,7 @@ class Request implements RequestInterface
     protected function isValidMediaTypeHeader($headerName)
     {
         $header = $this->getHeaderLine($headerName);
-        return strpos($header, "application/vnd.api+json") !== false;
+        return strpos($header, "application/vnd.api+json") === false || $header === "application/vnd.api+json";
     }
 
     protected function setIncludedFields()